requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dotenv==0.9.9
fastapi==0.116.1
h11==0.16.0
idna==3.10
mysql-connector-python==9.4.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
websockets==15.0.1
